// src/types/partner.ts
export interface IPartner {
    id:                    string;
    name:                  string;
    code:                  string;
    address:               string;
    email:                 string;
    phone:                 string;
    status:                string;
    startDateSubscription: number;
    endDateSubscription:   number;
    createdAt:             number;
    updatedAt:             number;
}